commit containing changes in uf-creator frontend

See merge request melodic/uf-creator!3

+++ Uf-Creator/src/main/java/eu/morphemic/ufcreator/function_creator/model/ByTemplateFunctionDTO.java
@@ -12,7 +12,7 @@ public class ByTemplateFunctionDTO {
     @JsonProperty("a")
     public double a;
     @JsonProperty("b")
-    public int b;
+    public double b;
     @JsonProperty("weight")
     public double weight;
     @JsonProperty("shape")
